requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
plotly>=5.24.0
watchdog>=4.0.0
xlsxwriter>=3.2.0
kaleido>=0.2.1
fpdf2>=2.7.9
pysnmp>=7.1.0
python-telegram-bot>=21.0,<22.0
python-dotenv>=1.0.1
